code refactoring: 1. taking out nested class - UserListPresenter 2. create package for class needs for mvp recycler view adapter
