--- app/src/main/java/com/example/popularlibraries/ui/base/ItemView.kt
@@ -1,5 +1,0 @@
-package com.example.popularlibraries.ui.base
-
-interface ItemView {
-    var pos: Int
-}
