+++ app/src/main/java/com/example/popularlibraries/ui/base/recyclerview/ItemView.kt
@@ -1,0 +1,5 @@
+package com.example.popularlibraries.ui.base.recyclerview
+
+interface ItemView {
+    var pos: Int
+}
