+++ app/src/main/java/com/example/popularlibraries/ui/base/recyclerview/ListPresenter.kt
@@ -1,4 +1,4 @@
-package com.example.popularlibraries.ui.base
+package com.example.popularlibraries.ui.base.recyclerview
 
 interface ListPresenter<V: ItemView> {
     var itemClickListener: (pos: Int) -> Unit
