--- app/src/main/java/com/example/popularlibraries/ui/users/view/UserItemView.kt
@@ -1,7 +1,0 @@
-package com.example.popularlibraries.ui.users.view
-
-import com.example.popularlibraries.ui.base.ItemView
-
-interface UserItemView:ItemView {
-    fun setLogin(login:String)
-}
