+++ app/src/main/java/com/example/popularlibraries/ui/users/view/recyclerview/UserItemView.kt
@@ -1,0 +1,7 @@
+package com.example.popularlibraries.ui.users.view.recyclerview
+
+import com.example.popularlibraries.ui.base.recyclerview.ItemView
+
+interface UserItemView: ItemView {
+    fun setLogin(login:String)
+}
